refactor(dot11): extract Dot11TaggedParameter into element module
